requirements.txt
fastapi[standard]==0.115.6
strawberry-graphql[fastapi]==0.253.1
pytest==8.3.4
flake8==7.1.1
black==24.10.0
requests==2.32.3
